+++ platforms/ide/tooling-api/src/main/java/org/gradle/tooling/events/problems/internal/DefaultOffsetInFileLocation.java
@@ -19,22 +19,22 @@ package org.gradle.tooling.events.problems.internal;
 import org.gradle.tooling.events.problems.OffsetInFileLocation;
 
 public class DefaultOffsetInFileLocation extends DefaultFileLocation implements OffsetInFileLocation {
-    private final long offset;
-    private final long length;
+    private final int offset;
+    private final int length;
 
-    public DefaultOffsetInFileLocation(String path, long offset, long length) {
+    public DefaultOffsetInFileLocation(String path, int offset, int length) {
         super(path);
         this.offset = offset;
         this.length = length;
     }
 
     @Override
-    public long getOffset() {
+    public int getOffset() {
         return offset;
     }
 
     @Override
-    public long getLength() {
+    public int getLength() {
         return length;
     }
 }
